# Family-Nutrition-Planner/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NutriNext / ParivarSehat AI — Family Nutrition Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /grocery-lists:
    get:
      operationId: listGroceryLists
      tags: [grocery]
      summary: Get grocery lists for a family
      parameters:
        - name: familyId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of grocery lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroceryList"
  /grocery-lists/generate:
    post:
      operationId: generateGroceryList
      tags: [grocery]
      summary: AI-generate a weekly grocery list with budget-optimized substitutions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateGroceryListBody"
      responses:
        "200":
          description: Generated grocery list with cheaper swaps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryList"
  /grocery/cheaper-alternative:
    get:
      operationId: getCheaperAlternative
      tags: [grocery]
      summary: DB-backed cheaper ingredient alternative lookup
      parameters:
        - name: item
          in: query
          required: true
          schema:
            type: string
          description: Ingredient name to find cheaper alternatives for
        - name: budget
          in: query
          schema:
            type: number
          description: Max cost per serving in INR (default 100)
        - name: diet
          in: query
          schema:
            type: string
            enum: [vegetarian, "non vegetarian", vegan, eggetarian]
          description: Filter by diet type
      responses:
        "200":
          description: Cheaper alternatives from recipe DB
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheaperAlternativeResponse"
        "400":
          description: item query parameter is required

  # ── Health Logs ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    GroceryItem:
      type: object
      properties:
        category:
          type: string
          description: "Vegetables|Fruits|Grains|Pulses|Dairy|Spices|Oil|Other"
        name:
          type: string
        nameHindi:
          type: string
        quantity:
          type: string
        estimatedCost:
          type: number
        cheaperAlternative:
          type: string
          nullable: true
        alternativeCost:
          type: number
          nullable: true
        priority:
          type: string
          enum: [essential, optional]
      required:
        - category
        - name
        - quantity
        - estimatedCost
        - priority
    GroceryListData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GroceryItem"
        totalEstimatedCost:
          type: number
        budgetStatus:
          type: string
          enum: [within, over, under]
        savingsTips:
          type: array
          items:
            type: string
        seasonalSuggestions:
          type: array
          items:
            type: string
    GroceryList:
      type: object
      properties:
        id:
          type: integer
        familyId:
          type: integer
        mealPlanId:
          type: integer
          nullable: true
        weekOf:
          type: string
          format: date
        items:
          $ref: "#/components/schemas/GroceryListData"
        totalEstimatedCost:
          type: number
        budgetStatus:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - familyId
        - weekOf
        - items
        - totalEstimatedCost
        - budgetStatus
    GenerateGroceryListBody:
      type: object
      properties:
        familyId:
          type: integer
        mealPlanId:
          type: integer
        pantryIngredients:
          type: array
          items:
            type: string
          description: List of ingredients already in pantry
        updateMode:
          type: string
          enum: [add, subtract]
          default: add
          description: subtract removes pantry items from grocery list server-side (deterministic)
      required:
        - familyId
    CheaperAlternativeResponse:
      type: object
      properties:
        item:
          type: string
        budget:
          type: number
        isFallbackYear:
          type: boolean
        alternatives:
          type: array
          items:
            type: object
            properties:
              recipeId:
                type: integer
              name:
                type: string
              nameHindi:
                type: string
                nullable: true
              cuisine:
                type: string
              diet:
                type: string
              costPerServing:
                type: number
              calories:
                type: number
              protein:
                type: number
              totalTimeMin:
                type: integer
              source:
                type: string
                enum: [recipe_db]
            required:
              - recipeId
              - name
              - costPerServing
              - source
        message:
          type: string
          nullable: true
      required:
        - item
        - alternatives

    # ── Health Logs ──────────────────────────────────────────────────────────
...